autenticacion del token en funciones expuestas

+++ servicios/actividades/actividades.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"strconv"
-	"strings"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/session"
@@ -44,23 +43,10 @@ func ConnectMongoDb(clientMongo *mongo.Client) {
 // insertar actividad
 func InsertActividad(c *fiber.Ctx) error {
 
-	//Obtencion de token
-	authHeader := c.Get("Authorization")
-	if authHeader == "" {
-		// El token no está presente
-		return fiber.NewError(fiber.StatusUnauthorized, "No se proporcionó un token de autenticación")
-	}
-
-	// Parsea el token
-	tokenString := strings.Replace(authHeader, "Bearer ", "", 1)
-
-	err, codigo := userGoogle.ValidacionDeUsuarioPropio(adminNotRequired, anyRol, tokenString)
-	if err != nil {
-		if codigo != "" {
-			codigoError, _ := strconv.Atoi(codigo)
-			return c.Status(codigoError).SendString(err.Error())
-		}
-		return c.Status(fiber.StatusNotFound).SendString(err.Error())
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
 	}
 
 	// obtencion de datos
@@ -101,6 +87,13 @@ func InsertActividad(c *fiber.Ctx) error {
 
 // obtener actividad por id
 func GetActividad(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("actividades")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	cursor, err := coll.Find(context.TODO(), bson.M{"idActividad": idNumber})
@@ -117,6 +110,13 @@ func GetActividad(c *fiber.Ctx) error {
 
 // obtener todas las actividades
 func GetActividadAll(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("actividades")
 	cursor, err := coll.Find(context.TODO(), bson.M{})
 	if err != nil {
@@ -131,6 +131,13 @@ func GetActividadAll(c *fiber.Ctx) error {
 
 // borrar actividad por id
 func DeleteActividad(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("actividades")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	result, err := coll.DeleteOne(context.TODO(), bson.M{"idActividad": idNumber})
