+++ servicios/novedades/novedades.go
@@ -120,23 +120,10 @@ func ConnectMongoDb(clientMongo *mongo.Client) {
 // insertar novedad
 func InsertNovedad(c *fiber.Ctx) error {
 
-	//Obtencion de token
-	authHeader := c.Get("Authorization")
-	if authHeader == "" {
-		// El token no está presente
-		return fiber.NewError(fiber.StatusUnauthorized, "No se proporcionó un token de autenticación")
-	}
-
-	// Parsea el token
-	tokenString := strings.Replace(authHeader, "Bearer ", "", 1)
-
-	err, codigo := userGoogle.ValidacionDeUsuarioPropio(adminNotRequired, anyRol, tokenString)
-	if err != nil {
-		if codigo != "" {
-			codigoError, _ := strconv.Atoi(codigo)
-			return c.Status(codigoError).SendString(err.Error())
-		}
-		return c.Status(fiber.StatusNotFound).SendString(err.Error())
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
 	}
 
 	//obtiene los datos
@@ -228,6 +215,13 @@ func InsertNovedad(c *fiber.Ctx) error {
 
 // obtener novedad por id
 func GetNovedades(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("novedades")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	cursor, err := coll.Find(context.TODO(), bson.M{"idSecuencial": idNumber})
@@ -246,6 +240,13 @@ func GetNovedades(c *fiber.Ctx) error {
 
 // Busqueda con parametros Novedades
 func GetNovedadFiltro(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("novedades")
 	var busqueda bson.M = bson.M{}
 	if c.Query("tipo") != "" {
@@ -298,6 +299,13 @@ func GetNovedadFiltro(c *fiber.Ctx) error {
 
 // obtener todas las novedades
 func GetNovedadesAll(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("novedades")
 	cursor, err := coll.Find(context.TODO(), bson.M{})
 	if err != nil {
@@ -315,6 +323,13 @@ func GetNovedadesAll(c *fiber.Ctx) error {
 
 // borrar novedad por id
 func DeleteNovedad(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("novedades")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	result, err := coll.DeleteOne(context.TODO(), bson.M{"idSecuencial": idNumber})
@@ -326,6 +341,13 @@ func DeleteNovedad(c *fiber.Ctx) error {
 }
 
 func UpdateEstadoNovedades(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	//se obtiene el id
 	idNumber, err := strconv.Atoi(c.Params("id"))
 	fmt.Println(idNumber)
@@ -368,6 +390,13 @@ func UpdateEstadoNovedades(c *fiber.Ctx) error {
 }
 
 func UpdateMotivoNovedades(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	novedad := new(Novedades)
 	if err := c.BodyParser(novedad); err != nil {
@@ -386,6 +415,13 @@ func UpdateMotivoNovedades(c *fiber.Ctx) error {
 }
 
 func GetFiles(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("novedades")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	var novedad Novedades
@@ -421,6 +457,13 @@ func GetFiles(c *fiber.Ctx) error {
 
 // ----Tipo Novedades----
 func GetTipoNovedad(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("tipoNovedad")
 	cursor, err := coll.Find(context.TODO(), bson.M{})
 	if err != nil {
@@ -437,6 +480,13 @@ func GetTipoNovedad(c *fiber.Ctx) error {
 // ----Cecos----
 // insertar cecos
 func InsertCecos(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	cecos := new(Cecos)
 	if err := c.BodyParser(cecos); err != nil {
 		return c.Status(503).SendString(err.Error())
@@ -470,6 +520,13 @@ func InsertCecos(c *fiber.Ctx) error {
 
 // obtener todos los cecos
 func GetCecosAll(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("centroDeCostos")
 	cursor, err := coll.Find(context.TODO(), bson.M{})
 	if err != nil {
@@ -485,6 +542,13 @@ func GetCecosAll(c *fiber.Ctx) error {
 
 // obtener los cecos por codigo
 func GetCecos(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("centroDeCostos")
 	idNumber, _ := strconv.Atoi(c.Params("id"))
 	cursor, err := coll.Find(context.TODO(), bson.M{"codigo": idNumber})
@@ -501,6 +565,13 @@ func GetCecos(c *fiber.Ctx) error {
 }
 
 func GetCecosFiltro(c *fiber.Ctx) error {
+
+	// validar el token
+	error, codigo, _ := userGoogle.Authorization(c.Get("Authorization"), adminNotRequired, anyRol)
+	if error != nil {
+		return c.Status(codigo).SendString(error.Error())
+	}
+
 	coll := client.Database("portalDeNovedades").Collection("centroDeCostos")
 	var busqueda bson.M = bson.M{}
 	if c.Query("descripcion") != "" {
